requirements.txt
# Requirements pinned from the project's virtual environment (generated)
# NOTE: System binaries required: Tesseract (for pytesseract), Poppler (for pdf2image), and Chrome/Chromium for `nodriver` scraping.
streamlit==1.47.1
pandas==2.3.1
numpy==2.2.6
plotly==6.3.1
openpyxl==3.1.5
beautifulsoup4==4.14.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.27.1
rapidfuzz==3.14.1
PyMuPDF==1.26.5
pytesseract==0.3.13
Pillow==11.3.0
pdf2image==1.17.0
nodriver==0.48.1
requests==2.32.4
